# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Alpha server built with Go and Chi",
        "title": "StayHealthy Swagger Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/plan/{id}": {
            "get": {
                "description": "Get diet macros",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "macros calculator"
                ],
                "summary": "Get diet macros",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid formatted ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/calculator.UserMacroDistribution"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "calculator.UserMacroDistribution": {
            "type": "object",
            "properties": {
                "activity": {
                    "type": "string"
                },
                "activity_description": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "bmr": {
                    "type": "integer"
                },
                "calories_distribution": {
                    "type": "string"
                },
                "calories_distribution_description": {
                    "type": "string"
                },
                "carbs": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "fats": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string"
                },
                "goal": {
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string",
                    "example": ""
                },
                "objective": {
                    "type": "string"
                },
                "objective_description": {
                    "type": "string"
                },
                "protein": {
                    "type": "integer"
                },
                "system": {
                    "type": "string"
                },
                "tdee": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
...
    }
}